add guide for 'upgrade by installing again'
In order to to update/upgrade to the newest release, you just install the newest version in a new conda environment by following the instructions above. Then you will have a new conda environment with the newest version of Qiime, along side the preexisting conda environment with the older version of Qiime.